requirements.txt
autopep8==1.4.4
asgiref==3.5.2
distlib==0.3.5
Django==4.0.6
django-ckeditor==6.4.2
django-countries==7.3.2
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
django-environ==0.9.0
django-multiselectfield==0.1.12
django-js-asset==2.0.0
Pillow==9.2.0
pep8==1.7.1
pycodestyle==2.5.0
pytz==2018.5
requests==2.28.1
Djheroku==0.6.1
filelock==3.7.1
humanize==4.3.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.2
